fix some bug for yolov5 case
