+++ training/benchmarks/yolov5/pytorch/config/_base.py
@@ -3,7 +3,8 @@ from typing import ClassVar
 
 # case info
 # chip vendor: nvidia, kunlunxin,  iluvatar, cambricon etc. key vendor is required.
-vendor: str = None
+# vendor: str = None
+vendor: str = "iluvatar"
 # model name
 # name: str = "MobileNetV2"
 name: str = "yolov5"
@@ -73,8 +74,7 @@ resume_from_checkpoint: bool = False
 device: str = None
 n_device: int = 1
 
-
-cfg: str = "/workspace/flagperf/training/benchmarks/yolov5/pytorch/model/yolov5s.yaml"
+cfg: str = "/workspace/flagperf/training/benchmarks/yolov5/pytorch/models/yolov5s.yaml"
 # hpy = "path/to/hpy.yaml"
 hyp: str = "/workspace/flagperf/training/benchmarks/yolov5/pytorch/dataloaders/hyp.scratch-low.yaml"
 
@@ -83,27 +83,45 @@ data: str = "coco.yaml"
 # cuda device, i.e. 0 or 0,1,2,3 or cpu
 device: str = "0"
 # initial weights path
-weight: str = "yolov5s.pt"
+weights: str = "/workspace/flagperf/training/benchmarks/yolov5/pytorch/models/yolov5s.pt"
 
 epoch: int = 300
 
-data_dir = "/data/sen.li/workspace/datasets/yolov5/coco/images"
-train_data = "train2017"
-eval_data = "val2017"
+# data_dir = "/data/sen.li/workspace/datasets/yolov5/coco/images"
+
+# Train/val/test sets as 1) dir: path/to/imgs, 2) file: path/to/imgs.txt, or 3) list: [path/to/imgs1, path/to/imgs2, ..]
+#path: ../datasets/coco  # dataset root dir
+data_dir: str = "/mnt/data/yolov5/train"
+train_data: str = "train2017.txt"  # train images (relative to 'path') 118287 images
+eval_data:str = "val2017.txt"  # val images (relative to 'path') 5000 images
+test_data:str = "test-dev2017.txt"  # 20288 of 40670 images, submit to https://competitions.codalab.org/competitions/20794
+
+# Classes
+nc: 80  # number of classes
+class_names: list = ['person', 'bicycle', 'car', 'motorcycle', 'airplane', 'bus', 'train', 'truck', 'boat', 'traffic light',
+        'fire hydrant', 'stop sign', 'parking meter', 'bench', 'bird', 'cat', 'dog', 'horse', 'sheep', 'cow',
+        'elephant', 'bear', 'zebra', 'giraffe', 'backpack', 'umbrella', 'handbag', 'tie', 'suitcase', 'frisbee',
+        'skis', 'snowboard', 'sports ball', 'kite', 'baseball bat', 'baseball glove', 'skateboard', 'surfboard',
+        'tennis racket', 'bottle', 'wine glass', 'cup', 'fork', 'knife', 'spoon', 'bowl', 'banana', 'apple',
+        'sandwich', 'orange', 'broccoli', 'carrot', 'hot dog', 'pizza', 'donut', 'cake', 'chair', 'couch',
+        'potted plant', 'bed', 'dining table', 'toilet', 'tv', 'laptop', 'mouse', 'remote', 'keyboard', 'cell phone',
+        'microwave', 'oven', 'toaster', 'sink', 'refrigerator', 'book', 'clock', 'vase', 'scissors', 'teddy bear',
+        'hair drier', 'toothbrush']  # class names
+
 imgsz = 640
 batch_size = 64
 
 data: str = "/workspace/flagperf/training/benchmarks/yolov5/pytorch/dataloaders/coco.yaml"
 gs = 32 # grid size (max stride)
-single_cls = True
+single_cls = False
 pad = 0.5
 augment = True
-cache = False
-rect = True
+cache = None
+rect = False
 
 rank = -1
 workers = 4
-image_weights = True
+image_weights = False
 quad = False
 prefix = ''
 shuffle = True
